+++ releaf-core/spec/helpers/application_helper_spec.rb
@@ -42,8 +42,7 @@ describe Releaf::ApplicationHelper do
     before do
       translation = Releaf::I18nDatabase::I18nEntry.create(key: "admin.xx.colors-red")
       translation.i18n_entry_translation.create(locale: "en", text: "Color red")
-      I18n.backend.backends.first.translations_cache = nil
-
+      Releaf::I18nDatabase::Backend.reset_cache
       allow(helper).to receive(:controller_scope_name).and_return("admin.xx")
     end