+++ releaf-i18n_database/spec/features/translations_spec.rb
@@ -194,7 +194,7 @@ feature "Translations" do
 
   describe "Lookup" do
     background do
-      I18n.backend.backends.first.translations_cache = nil # reset cache
+      Releaf::I18nDatabase::Backend.reset_cache
       allow( Releaf.application.config.i18n_database ).to receive(:auto_creation).and_return(true)
     end